Added user defined global variables
This addition will remove some of the code reuse between the different
hosts and will define a single place where some of the shared variables
will be defined.

Created a module that will be included by main configuration and
home-manager files(the top functions that include all the rest).

Important: this include is not optional because other modules use those
global variables.

By defining the hostname variable with a value, other global varieables
will be defined automatically by values that are derived from it.