Added Redshift to adjust the screen color temperature
This service will reduce the amount of blue light
emitted from the screen at night and by doing so will
reduce the strain on the eyes.

It uses the latitude and longitude to determine when
the night begins.

It can be combined with geoclue2 and get the location
from it but currently I decided to manually set the
latitude and longitude for it to use.

IMPORTANT:
=========
This service only works on with xorg
and when moving to wayland it will need to be replaced.
Currently gammastep look like a good alternative.
I have tested gammastep to work on xorg but was gettings
some error messeges(systemd log) so decided to use redshift
which does the job.