Removed direct setting of a network interface
I have removed a static IP for one of my interface
for the work PC. I instead will be using NetworkManager
settings, as this interface gets dynamically plugged
and unplugged constantly.

This also removed the issue with OS getting stuck on boot
if the NIC was not available at the time of the boot
for 1:30 minutes.