Added script for quick use of cheat.sh
This scrip is intended for the use in a shell
and for a very convenient flow inside of a tmux
session.

By pressing a tmux key binding a new window will open
for the user to interact with cheat.sh via the script.
Once user is done with the information he will press enter
at which point the tmux window will get closed and the user
will be back at the window he were when he pressed the
key binding to start his cheat.sh query.