Fixed moolticuted not working after reboot
When PC started the moolticuted systemd service would not work
properly(browser add on was red and was not communicated with the
connected and unlocked mooltipass device), and only started working
once systemd service was restarted or mooltipass device locked and
unlocked.

By adding "after" section to the moolticuted systemd file I was
able to make it work. It looks like the service was starting
before all the other services it depended on.

Unfortunately even after this change there still was some race condition
and I add to add a sleep of 5 seconds to fix the issue. So right now
this is a working work around.

Also made "description" a valid NixOS syntax.