Updated the script used to open Firefox bookmarks with rofi
Script worked just fine when Firefox was in i3wm active work space
but as soon as it was not, the link would open just fine but the user
needed to switch to the work space with Firefox in it manually to see
the opened bookmark.

This enhancement makes i3wm switch to the work space automatticly with
Firefox in it even if this work space is currently isn't not the
active one.

This enhancement is only relevant to the i3wm and uses i3-msg.