Refactored moolticuted
Logically udev rules shouldn't be grouped together, as each one is specific
to some device and not all rules are required by hosts that may include
the udev.nix file. And so udev.nix was removed.

Some devices might have a companion app that needs more then a daemon
running but also some udev rule. A good example is the moolticuted.
So in such case it makes sense to couple the udev rule and the systemd
service(daemon) into a single .nix file as the two essentially a single
unit.