Tmux will now automatically restore saved session on boot
- Added directives to continuum and restore tmux plugins so
when tmux server starts it will automatically restore(using continuum
plugin) the sessions that were saved(using restore plugin).
- Fixed systemd tmux unit to start tmux with "start-server" switch so
when started no new sessions will be created(as it was when using
"new-session -d" switches) This way tmux will be start on boot
it will restore saved sessions any nothing else.

Note: the systemd aspect of tmux is defined in a seperate .nix file
and not generated by the continuum plugin settings, as this is NixOS
and systemd units hare handled as nix setting files.