Partial fix tmux bindings.
I have moved the builtin copy-yank keybindings of tmux from the
tmux-yank plugin and made sure the do not conflict with the key
bindings that are used by the plugin.

I still need to make sure if I need the builtin keybindings, with
the functionality provided by the tmux-yank being superior.