tmux-yank plugin comment
The copy buffer for tmux is separate from the system one, in the past in
order to sync the two there was a need to install tmux-yank but it looks like
Tmux now sends the OSC52 escape code that tells the terminal(one that support this)
to not display the following characters, but to copy them into the clipboard
instead.

So when testing with my terminal emulator alacritty I noticed that indeed there was no need
in an extra plugin to sync the two clip boards.

Note that not all the Tmux plugins support the use of OSC52 escape,
which means that in the future there might be a need to create
workarounds for them or the default to installing tmux-yank. Also this
"native" copy sync might break when moving to a different terminal emulator.

The reason that this plugin is still included because it provides a quick way to copy what
what is on the command line and once in copy mode to copy the PWD. I might just replace the
plugin with keybindings(based on send-keys).