Added tmux-thumbs plugin
Only parses the strings on screen and not the entire "shell buffer" and
filters things like File paths, Git SHAs, IPv4, IPv6 addresses...
It then assigns to each of the relevant strings a key that you can
press and it will copy the string, this is how vimium/vimperato works ass well.

This is a very fast and convenient way to copy relevant strings.