Added extrakto tmux plugin
This plugin is a fast way to copy text from the tmux-copy mode using
fuzzy finder and once selected copy them to the clip board which is much
faster then going via the tmux built in copy mode(even when using search
in it). As a bonus there is also a way to open links and files directly
from the fuzzy finder menu.

This commit also adds user defined keybindings for extrakto and
tmux-thumbs plugins.