tmux-resurrect plugin will now restore applications
I have fixed the issue causing tmux-resurrect to fail resurrecting
applications.

The strings that are saved for some applications such as nvim, vim,
man... when using NixOS, appimage, asdf-vm into the tmux_resurrect_* files
can't be parsed and restored. I have leveraged a hook that is provided
by the tmux-resurrect plugin in order to execute a sed command on the
tmux_resurrect_* file when it is created and fix the strings causing
the issues in mux_resurrect_* files so they can be parsed by the
tmux-resurrect plugin and successfully resurrected.

I have also tweaked the tmux automatic saving of state to 10 minutes
and updated the README.