Updated i3wm keybindings for quick tmux access
I have added a keybinding that will start a new terminal and there
it will attach to tmux existing session or create a new one if there is
no existing sessions.

The command I used for this is on par with the command that opens
a new terminal that starts zsh shell but in order to start tmux
I needed to be explicit about how my alacritty terminal starts,
I stated what shell to start(-e), in what mode(-i interactive)
and finally what command to execute(-c).