Added vterm and eshell settings
Have added a very basic settings for vterm and the eshell
with one of the cosmetic settings for eshell not working and need
to be looked into.

This is a very basic setup and if I decide to use my shells via emacs
I will need to give it more look and time.