Added rainbow-delimiters-nvim plugin to nvim
This plugin is currently not available on the stable branch
of nixos and so I created a package definition for it.

I also configured the order of the colors.
As of now there are some languages(such as cpp) that this plugin
is not working for me. I am not sure why yet.