Fixed ctags issue introduced by installing Emacs
Emacs has its own binary for ctags and it got picked up over the
universal-ctags package I was using before.

This fix places the universal-ctags to be listed before Emacs package
and by doing so solves the issue.

I am not sure if this is the nix idiomatic way to go about this
issue.