Refactoring starship prompt
In order to create a more informative command line prompt
I refactored the settings for starship prompt and provided it
a look that reminds the "lean style" used by Powerlevel10k.

I decided to stick with starship as it is not limited to zsh
and can be reused with other shells such as bash which is the
"default" when it comes to remote servers.