Added SSH_AUTH_SOCK to watchman service
SSH_AUTH_SOCK is required for executing commands that involve
ssh such as rsync. When using rsync in the context of systemd
there is no option to enter a password and so access to ssh-agent
is required. The path to the socket needs to be explicitly stated as
it is not included in the "setEnvironment".

The string that is used as a value for SSH_AUTH_SOCK is NixOS
and will need to be changed for other OS, plus I am using
gpg-agent as a ssh-agnet.