Refactored flake.nix
I have removed some code that were reused(nixosConfigurations) and
more importantly made the flake ready to be used with a stand alone
installation of home-manger, which is useful in case where NixOS is not
an option.

This is an initial step and mostly makes sure that the $HOME flake
defined settings can be evaluated and applied using a stand alone home-manger
installation. This does not yet cover 1:1 the experience that a user will get
using the flake with NixOS.