Defined global variable for definition of font
I have defined a variable to have the value of a font
that can then be used all over the system to eliminate code reuse
and to create a uniform font across the entire system.

I have updated alacritty, i3, to name a few to use this global variable.
For this variable to work the font family(Nerd font) need to be
installed. cli software such as vim and tmux use terminal emulator for
there fonts.