Moved nvim-treesitter plugin to use declarative style for its grammars
I removed the setting for the plugin to install its grammars if they do
not exist in a predefined directory and instead replaced this with a nix
style predefined grammars.

Hopefully this will fix an issue I was having where treesitter failed to
read its grammars that it installed in "run time".