Updated starship prompt
Did a small change to the order of the modules. I put the status
module before before time and cmd_duration. This way time and
cmd_duration are always displayed at the end of the prompt with
status module getting "appended" to them when it is needed and the
loom and fill stays consistent regardless if status is displayed or not.