Updated tmux-sessionizer.sh to be less depended
In a regular bash script if some software is messing
the script will fail to execute. As the script is being
generated using nix I can make sure the script will install the
messing software when it will be installed and then software
will removed with it and picked up by the garbage collector.