Created additional disko configurations for mount only hard drives
As disko used both for formatting hard drives and then mounting them
as part of the NixOS configuration. The hard drives that are only
to be mounted need to be broken off into a separate file file for NixOS
to import. If the separation is not done the disko "format" command
will format them as well.