Fixed tmux resurrect-continuum plugins edge case on a fresh install
tmux continuum plugin was only working after user manually saved
the first session using the resurrect plugin and now this is no longer
an issue.

Also refactored the directory to keep the resurrect session files
and defined it to be in the same root at the config for tmux.
which also aligns with the XDG base directory specification.