Updated i3wm keybindings
I removed the explicit nix path from i3wm for the terminal,
because terminal emulators such as alacrity use GPU acceleration
and this requires additional work when home-manger is used
as a stand alone module on a generic destibution such as Ubuntu.

So the solution, for now is to install alacrity using apt
and no the nix package manger and by not epecifing explicit
path to it in the configuration for i3wm the alacrity that
will be picked is the native(to Ubuntu) one that was installed
using apt.