Temporary fix for nix-index
nix-index is used for command-not-found when using flake.
I was using an option where I used pre compiled index as a file inside
of my home directory. This option got broken in the latest nix-index
update. For the time being there is a need to execute nix-index from the
cli for the index to work.

I will look into the recompiled index file option if it is still
possible.