Added "support" for alacritty terminal to display images
The double quotes because this support is in a form of ueberzugpp app
that other apps running in the terminal can use in order to draw images
inside of alacritty.

This is unlike a terminal such as kitty where is already has a component
that app can use to draw.