Fixed an emacs issue where upon saving buffer cursor position will change
When saving in normal mode or insert mode(emacs evil) on a new heading
my cursor would "jump up". I was able to pinpoint the issue with the
toc-org plugin that maintenance dynamic list of links to headings in the
org-mode document, this list will gain a new entry each time a new
heading was created which "pushed the page down" and created of the
effect of the cursor "jumping up"

I also suspect that me defining some constant page margins might also has some
effect. Because prior to this fix, When I removed the use of plugin, my
cursor stopped moving but the line on which I was "moved" on the
monitor, it looks like this fix address both.