Updated bootstrap script to copy age keys
To support sops-nix a key is required during initial installation
this commit updates the bootstrap script to copy it to the expected
path if key exists.

This is initial commit to address sops-nix support as part of fresh
installation of NixOS on a new machine. And more design is required
to streamline it.