Improved misspelling flow in neovim
I was able to jump to misspelled words using ]s and ]s and once
word was fixed the cursor would go to the start of the word I was
fixing.

This fix makes sure that after the word is fixed the cursor will
be placed at the end of the word.