Streamlined command_not_found for flake and now it is pre-generated
By default command_not_found only works with non flake based
configurations. For flakes the alternative is to use nix-index,
but this option requires to manually generate the database for the
command. The better approach is to use nix-index-database which is a
wrapper around nix-index with pre-generated database.

This commit removed all traces of nix-index and moved completely to
the use of nix-index-database flake input. As suggested by the
nix-index-database documentation in any case you need to pick nix or
nix-index-database as they might conflict.