Fixed custom dictionary for neovim on new machine
As I am using home-manger defined dictionary for neovim it is not
enough to create the .add file at ~/.config/nvim/spell but there
is also a need to create from it a binary file(.spl). This is
done using a neovim command.

I have fixed my configuration to create this file if it does not exist
already, addressing the edge case of a new machine setup.