Added keybindings to scroll the preview of lazygit
The goal was to remove the friction out of going from the keyboard to
the mouse and back each time I wanted to scroll the preview of a search
result.

I am not 100% happy with the result, and there are some limitations on
on what I could bind especially when using lazygit inside of tmux
session.