Added keybindings to scroll neovim search results preview
The goal was to remove the friction out of going from the keyboard to
the mouse and back each time I wanted to scroll the preview of a search
result.

Added the use keybindings for scrolling the preview that is provided by
telescope and fzf-lua plugins.

As my arrow keys are set on my hjkl as a separate keyboard layer I
can use j/k to scroll up and down while holding my other layer.