Fixed emacs org-roam leaving images after node is deleted
org-roma is a great tool to create nodes that are wrappers
around org files. Wrapper that can be searched and have meta data.

org-roma provides very basic functionality to create new nodes but
it does not provides an options to delete nodes and to rename them.

Both of which I already implemented in the past, instead of doing
this manually.

As I rarely work with images I did not noticed that although my
implementation handles the removal of nodes it does not remove the
images inside of them that are just files inside of the roam directory
located all around it.

This patch does two things
- when images are inserted using org-download plugin a directory is
  created inside of the <root of roam directory>/images. This directory
  uses the unique number that every org-roam gets as a prefix to its
  name(E.g: ../roam_notes/images/13102024211520).
- When org-roam node is deleted, then this unique directory is deleted
  as well removing all the associated images with the node.

By using only the unique prefix and not the entire name of the node
the path to the images inside the org documents stay valid even if
the name of the node gets updated(by yet another function that I
implemented).