Refactoring of emacs configuration
I have refactored the configuration based on responsibility into
separate files. As the current state was becoming difficult to work
with.

Using use-package:
- I took tighter control over the packages by :after options.
- As my package manger is nix I explicitly set :ensure to nil for
all the packages so there will be no case in which use-package
will try to download packages and create potential conflicts.
- Better encapsulation of configuration into the user-package stricture.

Fixed some variables that were not being set.
Moved form using relative paths.

Removed two themes that were not being used.

Started breaking parts of general package configuration into the
packages that are using it. This way it will not have undefined
keybindings when some package get removed in the future.