Refactored into host-specification
The main goal was to move away from meta.nix which couples global
variables and logic that is based on an if statements.

As a result I have created host-specification.nix file that holds a
scheme of variables to be defined by each hosts(username...).

I also took this opportunity and not only cleaned up meta.nix but
finally stopped importing it twice(once for nixos configuration and a
second time for home-manager that is used as nixos module). This commit
includes both meta.nix and host-specification.nix once and then the
values are passed as a variable to home-manager to use.

Note: right now home-manager as a stand alone module is not yet working.