Refactored meta.nix global options and removed if flow control
The meta.nix had all of the userDefinedGlobalVariables which were used
across all of the configurations(nixos and home-manager) regardless of
the actual scope of the of such options. With the main effort to encapsulate
configurations and couple global options into the relevant modules.

- Some options were replaced with static module configurations(let cfg = .. in),
  with those options mainly used to eliminate DRY inside the module.
- Some options were moved inside of the relevant module, so they can
  provide value inside the module and expose the option as API for other
  modules to use.
- Some options were removed completely(like port related options), as
  exciting module options can be used instead directly.

The second issue was the if flow control that was setting bunch of
options based on hard coded hostName provide by specialArgs/extraSpecialArgs

The use of hostName by itself was fine but I didn't like the fact that
I was mixing state and logic, and it made much more sense(scaling wise
as well) to remove the if statement and instead set the options
inside of each machine's  high level configuration files.