Refactored repository into "import all and enable" pattern
Each machine has main high level configuration files.
- home.nix
- configuration.nix
- configuration-services.nix
- hardware-configuration.nix

Each such file imports bunch of other files, each representing a low
level module(E.g, neovim, fzf...). This creates a lot of duplication, E.g adding
new module file means editing all of the imports list for all of the machines,
which does not scale well at all and creates overhead and the opportunity to
miss a machine, which is unlike changes that are made inside of module file that
acts as a single source of truth and effects all of the machines that
are importing it.

This commit seeks to solve this by "importing all and enable" pattern.
All of the low level modules are imported via the high level
configuration files and each such configuration file can then enable
only the low level modules it needs to use.

A step farther is to create "profile options" that aggregate the low level module enable
options, which allows for all "desktops" / "servers" to "flick a single switch" at the
high level configuration files and get bunch of modules enabled at once.
This way adding a low level module to one of the profile will enable it
for all of the machines using the profile.

To achieve this:

- All of the individual low level modules are wrapped inside an if
  statement that is activated by a conditional variable that is added to
  the module and is exposed to higher level modules to use.
- At the root of each directory with the low level modules(E.g
  modules/home-manager, modules/nixos...) a default.nix file was created
  that is used to import all of the low level modules in that directory.
- Indie each such default.nix file another option was created that is
  used to aggregate the low level modules by profile(gaming, desktop,
  server..).
- Each high level configuration file imports the default.nix file and
  enables the desired profiles for it to use. Low level modules that
  are not part of a profile are enabled individually.

I also used this opportunity to clean up modules such as syncthing and to
rename and move some of the files to better convey their intent.