Created host for mining crypto
This is the first step in creating a machine that will be dedicated
to mine monero. The services that I will need for mining will
be done next.

The goal is to use the decentralized p2pool that will be running
locally on mining machine, it will require access to the monero block
chain, which will be provided by monerod(running in --prune-blockchain
to save storage). p2pool will be giving jobs to the xmrig miner that
will be doing the actual work.

For now the full setup was only tested on VM(mining flow, monerod,
p2pool, xmrig were tested on actual hardware and are working just fine).
Once I deploy this configuration on an actual hardware I will need to
update disko.