Set a valid locale for the system
The GUI installer I used a long time a go used the location to
set locale settings, The issue was with the locale it picked that did
not set valid values for variables such as LANG, which caused me
to export them as part of the shell rc file.

Without such valid files tmux was unable to display valid UTF8
characters.

I have set en_US.UTF-8 as a solid base and I am using extraLocaleSettings
for specific variables configurations.